requirements.txt
batchgenerators==0.24
comet_ml==3.0.2
dotmap==1.3.30
imageio==2.21.1
matplotlib==3.5.3
numpy==1.22.3
opencv_python==3.4.18.65
oyaml==0.9
pandas==1.4.3
Pillow==9.5.0
PyYAML==6.0
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.9.0
seaborn==0.11.2
skimage==0.0
torch==1.12.1
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.64.0
zennit==0.5.1
